add readConfig function
This function handles netrc reading, remove that from the bin code.

This commit also removes support for the deprecated `url` field, we've
gone to 2.0.0, it's fine to remove the support for it now.
