+++ client/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "gooi",
-	"version": "2.0.0",
+	"version": "2.1.0",
 	"description": "Transfer files with gooi",
 	"main": "main.js",
 	"scripts": {
